pyproject.toml
[tool.poetry]
name = "desafio-backend"
version = "0.1.0"
description = ""
authors = ["Pedro Fonseca <contatopedrorn@gmail.com>"]
readme = "README.md"
packages = [{include = "desafio_backend"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.5"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
ruff = "^0.0.252"
pre-commit = "^3.1.0"
coverage = "^7.2.1"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.per-file-ignores]
"src/imo_banco/settings.py" = ["E501"]
"src/app/migrations/0001_initial.py" = ["E501"]
"src/app/domain/models/__init__.py" = ["F401"]
"src/app/models.py" = ["F401"]
"src/app/signals.py" = ["F401"]
"src/app/apps.py" = ["F401"]
